Change archive function to complete, change delete to archive

+++ src/containers/TodoListContainer/TodoListContainer.tsx
@@ -16,26 +16,24 @@ import styles from "./TodoListContainer.module.scss";
 
 const TodoListContainer = () => {
   const [todos, setTodos] = useState<TodoResponse[]>([]);
-  const [currentTodos, setCurrentTodos] = useState<TodoResponse[]>([]);
-  const [archivedTodos, setArchivedTodos] = useState<TodoResponse[]>([]);
   const [categories, setCategories] = useState<CategoryResponse[]>([]);
+  const [error, setError] = useState<Error | null>(null);
 
   const onTodoSubmit = async (data: TodoFormData) => {
     createTodo(data)
       .then((data) => {
-        const updatedTodos = todos.filter((todo) => todo.archived !== true);
-        setCurrentTodos([data, ...updatedTodos]);
+        setTodos([data, ...todos]);
       })
-      .catch((e) => console.log(e));
+      .catch((e) => setError(e));
   };
 
-  const onTodoComplete = (data: TodoResponse) => {
-    if (data.archived) {
-      const archivedTodos = todos.filter((todo) => todo.archived == true);
-      setArchivedTodos([...archivedTodos, data]);
-    } else {
-      setCurrentTodos([data, ...todos]);
-    }
+  const onTodoComplete = () => {
+    console.log("Todo complete!");
+    getAllTodos()
+      .then((data) => {
+        setTodos(data);
+      })
+      .catch((error) => setError(error));
   };
 
   useEffect(() => {
@@ -43,49 +41,34 @@ const TodoListContainer = () => {
       .then((data) => {
         setTodos(data);
       })
-      .catch((error) => console.log(error));
+      .catch((error) => setError(error));
     getAllCategories()
       .then((data) => setCategories(data))
-      .catch((error) => console.log(error));
+      .catch((error) => setError(error));
   }, []);
 
-  useEffect(() => {
-    setCurrentTodos(todos.filter((todo) => todo.archived !== true));
-    setArchivedTodos(todos.filter((todo) => todo.archived == true));
-  }, [todos]);
-
-  const onDelete = async (id: number) => {
-    const isDeleted = await deleteTodoById(id).catch((e: any) => {
-      console.log(e);
-      return false;
-    });
-    if (isDeleted) {
-      const currentTodos = todos.filter((todo) => todo.id !== id);
-      setTodos(currentTodos);
+  const onDelete = async (id: number, completed: boolean) => {
+    if (!completed && !confirm("Delete incomplete task?")) {
+      return;
+    } else {
+      const isDeleted = await deleteTodoById(id).catch((e: any) => {
+        setError(e);
+        return false;
+      });
+      if (isDeleted) {
+        const updatedTodos = todos.filter((todo) => todo.id !== id);
+        setTodos(updatedTodos);
+      }
     }
   };
 
-  // console.log("current render");
-  // console.log(currentTodos);
-  // console.log(archivedTodos);
-
   return (
     <>
-      <h3>{!todos.length && "Relax! You're all caught up."}</h3>
+      {error && <h1 className={styles.Error}>Error: {error.message}</h1>}
       <div className={styles.TodoListContainer}>
         <TodoForm onTodoSubmit={onTodoSubmit} categories={categories} />
-        {currentTodos.map((todo) => (
-          <TodoCard
-            key={todo.id}
-            todo={todo}
-            onTodoComplete={onTodoComplete}
-            onDelete={onDelete}
-            categories={categories}
-          />
-        ))}
-      </div>
-      <div className={styles.archiveContainer}>
-        {archivedTodos.map((todo) => (
+        {!todos.length && !error && <h3>Relax! You're all caught up.</h3>}
+        {todos.map((todo) => (
           <TodoCard
             key={todo.id}
             todo={todo}
